+++ CHANGELOG.md
@@ -29,3 +29,4 @@ to structure this file.
 ## [Unreleased]
 
 - Handle null, true, and false values in default testcases.
+- Add option to load notebook instead of python file.