+++ CHANGELOG.md
@@ -34,3 +34,4 @@ to structure this file.
 - Add support for multiple methods in solution class.
 - Add command to open description for an existing template.
 - Add an error message for invalid cookie.
+- Add a configuration to save default study plan.